feat(suite): show zero balance tokens
